Crypto Book SSE endpoint implementation
