+++ iextrading4j-client/src/test/java/pl/zankowski/iextrading4j/client/sse/request/alternative/CryptoBookSseRequestBuilderTest.java
@@ -1,7 +1,7 @@
 package pl.zankowski.iextrading4j.client.sse.request.alternative;
 
 import org.junit.Test;
-import pl.zankowski.iextrading4j.api.alternative.CryptoDetailedBook;
+import pl.zankowski.iextrading4j.api.alternative.CryptoBookEvent;
 import pl.zankowski.iextrading4j.client.sse.manager.SseRequest;
 
 import javax.ws.rs.core.GenericType;
@@ -16,12 +16,12 @@ public class CryptoBookSseRequestBuilderTest {
     public void shouldSuccessfullyCreateRequest() {
         final String symbol = "IBM";
 
-        final SseRequest<List<CryptoDetailedBook>> request = new CryptoBookSseRequestBuilder()
+        final SseRequest<List<CryptoBookEvent>> request = new CryptoBookSseRequestBuilder()
                 .withSymbol(symbol)
                 .build();
 
         assertThat(request.getPath()).isEqualTo("/cryptoBook");
-        assertThat(request.getResponseType()).isEqualTo(new GenericType<List<CryptoDetailedBook>>() {});
+        assertThat(request.getResponseType()).isEqualTo(new GenericType<List<CryptoBookEvent>>() {});
         assertThat(request.getPathParams()).isEmpty();
         assertThat(request.getQueryParams()).contains(entry("nosnapshot", "false"), entry("symbols", symbol));
     }
